server/requirements.txt
numpy
pandas
google-cloud-bigquery
db-dtypes # for GBQ
functions-framework==3.* # For cloud functions